+++ compiler/arossupport/include/libcall.h
@@ -70,11 +70,11 @@ typedef void (*VOID_FUNC)();
 #endif
 #ifndef __typedef_LONG_FUNC
 #define __typedef_LONG_FUNC
-typedef long (*LONG_FUNC)();
+typedef int (*LONG_FUNC)();
 #endif
 #ifndef __typedef_ULONG_FUNC
 #define __typedef_ULONG_FUNC
-typedef unsigned long (*ULONG_FUNC)();
+typedef unsigned int (*ULONG_FUNC)();
 #endif
 
 /* Declare all macros which the systems' libcall didn't */
